Implement circular buffer queue (will revert next)
Replaced the simple queue{end+1}/queue(1)=[] with a circular buffer to match how the MEX code does it. Pre-allocate the queue and track head/tail positions with wrap-around instead of shifting arrays around.

I get around 2% speedup (tested on small/deep/wide networks) but the code gets messier - now tracking 3 variables
  (qHead, qTail, qSize) with wrap-around logic instead of just queue{end+1}.

Reverting this in the next commit because the simple version is way easier to read and maintain. The ~2% gain isn't worth it when the real bottleneck is the transfer functions anyway. Keeping this commit in history in case we need it later.